requirements.txt
lxml>=3.6.4
python-dateutil>=2.5.3
